Refactor: 상품 DTO 수정
product 엔티티를 DTO로 변환하는 과정을 service단에서 처리하지 않고
DTO 안에서 처리하도록 수정했다.
